Add DeleteRowButton component
